// src/src/types.ts
export interface Product {
  id: string;
  name: string;
  price: number;
  stock: number;
  category: string;
  image_url?: string;
  description?: string;
}